% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aevent2df.R
\name{aevent2df}
\alias{aevent2df}
\title{Cladogenetic Event table to summary data frame}
\usage{
aevent2df(event_tables, type = NULL)
}
\arguments{
\item{event_tables}{a list with anagenetic tables acquired through the stochastic mapping from BioGeoBEARS.}

\item{type}{type of anagenetic event, NULL by default. This can be subset dispersal "d" and extinction "e".}
}
\value{
a data.frame with the type of event, the time at which it happened in MYA, at which node it happen and the area in which it happened.
}
\description{
Transfors the standard anagenetic event tables from BioGeoBEARS SBMs into a data.frame that includes the anagenetic event of interest across the total number of calculated maps.
}
