% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus2lemad.R
\name{clus2lemad}
\alias{clus2lemad}
\title{Cluster to Lemad}
\usage{
clus2lemad(data, tree)
}
\arguments{
\item{data}{A data.frame with species and a unique range (e.g., "A", "B"), in that order.}

\item{tree}{A phylogenetic tree of the species from data.}
}
\value{
A data frame of species and their combined ranges in the same order as the phylogeny.
}
\description{
This function transforms a data frame of species with their ranges into a working table ready to use for Lemad analysis. By combining all the individual ranges and sorting the species in the same order as they are in the phylogeny.
}
