% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event2df.R
\name{event2df}
\alias{event2df}
\title{Cladogenetic Event table to summary data frame}
\usage{
event2df(event_tables, type = NULL)
}
\arguments{
\item{event_tables}{a list with cladogenetic tables acquired through the stochastic mapping from BioGeoBEARS.}

\item{type}{type of cladogenetic event, NULL by default. This can be subset cladogenesis "subset (s)", vicariance "vicariance (v)", in situ speciation "sympatry (y)", and founder events "founder (j)".}
}
\value{
a data.frame with the type of event, the time at which it happened in MYA, at which node it happen and the area in which it happened.
}
\description{
Transfors the standard cladogenetic event tables from BioGeoBEARS SBMs into a dataframe that includes the cladogenetic event of interest across the total number of calculated maps.
}
