% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam2points.R
\name{pam2points}
\alias{pam2points}
\title{PAM to points}
\usage{
pam2points(PAM, phylo_names = F, gridid = F)
}
\arguments{
\item{PAM}{a Presence Absence Matrix object from letsR.}

\item{phylo_names}{option to turn the names of species into names that can match phylogenetic data.}

\item{gridid}{option to return the id of the grid, useful for validate number of assamblages.}
}
\description{
Transforms a Presence Absence Matrix object from letsR into geographic data points for the species.
}
