% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res2df.R
\name{res2df}
\alias{res2df}
\title{BioGeoBEARS result to data.frame}
\usage{
res2df(res, phylo, areas = NULL)
}
\arguments{
\item{res}{res object (calc_loglig_sp_results) from BioGeoBEARS}

\item{phylo}{phylogenetic tree used to calculate res object}

\item{areas}{number of unique areas used to calculate res object}
}
\value{
a data.frame with the tip label and node number in the phylogey and the most probable ancestral area fro such label.
}
\description{
This function helps extracting the state with maximum likelihood probability for each branch at the node (i.e., the most probable ancestral area for each node, and tip). This is useful for plotting in ggtree or better visualizing the ancestral states of the node.
}
