% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2cluster.R
\name{sp2cluster}
\alias{sp2cluster}
\title{Species to clusters
This function assigns species to the clusters obtained with phyloregion. Gives which species live in which clusters}
\usage{
sp2cluster(comm_data, cluster_data, letters = T)
}
\arguments{
\item{comm_data}{A sparse community data from phyloregion}

\item{cluster_data}{A phyloregion data.frame result}

\item{letters}{This function allows the user to decide if the clusters will be given as numbers (phyloregion output) or as letters (suitable for BioGeoBEARS, Lemad or GeoHiSSE)}
}
\description{
Species to clusters
This function assigns species to the clusters obtained with phyloregion. Gives which species live in which clusters
}
