% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2cluster.R
\name{sp2cluster}
\alias{sp2cluster}
\title{Species to clusters
This function assigns species to the clusters obtained with phyloregion. Gives which species live in which clusters}
\usage{
sp2cluster(comm_data, cluster_data, letters = T)
}
\arguments{
\item{comm_data}{a sparse community data from phyloregion}

\item{cluster_data}{a phyloregion data.frame result}
}
\description{
Species to clusters
This function assigns species to the clusters obtained with phyloregion. Gives which species live in which clusters
}
